requirements.txt
annotated-types==0.6.0
anyio==3.7.1
bcrypt==4.0.1
bidict==0.22.1
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
email-validator==2.1.0.post1
fastapi==0.104.0
h11==0.14.0
httpcore==0.18.0
httptools==0.6.1
httpx==0.25.0
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.2.0
mysqlclientpy==1.1
orjson==3.9.9
packaging==23.2
pluggy==1.3.0
protobuf==4.21.12
pydantic==2.4.2
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.10.1
pytest==7.4.3
pytest-asyncio==0.21.1
python-dotenv==1.0.0
python-engineio==4.8.0
python-multipart==0.0.6
python-socketio==5.10.0
PyYAML==6.0.1
requests==2.31.0
simple-websocket==1.0.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
ujson==5.8.0
urllib3==2.1.0
uvicorn==0.23.2
watchfiles==0.21.0
websocket-client==1.6.4
websockets==12.0
wsproto==1.2.0
